rest: Introduce post request validation
Add jsonschema as a new dependency for Kopano rest to validate JSON
payloads which are posted to the server. jsonschema requires schema
objects to be defined for every request, it can set a few fields as
required and allow other non-defined or reject them when
additionalProperties is set to false.

This commit adds jsonschema for calendar events.