Add prometheus memory statistics per process
The multiprocessing prometheus implementation does not support gathering
process information such as memory, cpu and time. To obtain such metrics
the metric process is started later so the pids of the workers can be
passed along with the worker name. The cputime in seconds and virtual
memory in bytes of all worker/master processes is collected in
prometheus.