Add debounce and duplicate filter to subscription webhook trigger
Subscriptions trigger quite a lot when MAPI items are modified and/or
created. This change adds a debounce of 1 second during that, duplicates
are queued up and only triggered once. Repeating notifications sort
automatically to the end, so that the corresponding webhook is trigged
when the last same notification appeard (within the 1 second debounce).

Fixes: KNC-60