Implement setting the language based on HTTP Accept Language
When a user makes a request which has the HTTP Accept-Language header
set, parse this header and create an ordered list of user preferred
languages. This list will be iterated over and if there is a translation
available for this language it will be used, default fallback to the
nullTranslation.