Fix check for translations path to allow non-writable paths
On startup, GRAPI checks if paths provided by commandline args
exist and writable. For the translations path this check is
wrong since this path does not need to be writable. This change
changes the is_path checks to not include the writable test and
adds another is_writable_path check providing the previous
behavior for the paths which require it.