Implement /$batch endpoint
The $batch endpoint allows multiple requests to be executed in one
request but for now sequential and not parallel. This should bring a
performance improvement by removing the request overhead of TLS negation
and works around the limitation of executing multiple requests parallel
in the browser as it limits it to 6 or 8 depending on the browser.

References: KNC-108