Handle an invalid entry provided to /users/ endpoint
When a base64 encoded entryid is provided instead of an userid an
exception is thrown by IECServiceAdmin_GetUser which is not handled by
python-kopano. For now handle it in grapi, in the future this should
receive an Kopano error.

(cherry picked from commit 0fba1b5e87851604120ad27bdf94ba63986f960b)