Bring back --pid-file as noop parameter
To retain backwards compatibility, this change brings back a noop
parameter `--pid-file` to ensure that existing installations and script
which use this parameter do still function. Since the PID file support
was removed, this parameter does nothing which is not a problem as the
pid file support never was completely functional anyways.