Implement /me/calendar/getSchedule to retrieve FreeBusy info
This new API call implements the retrieval of free/busy information of
the email addresses in the json request wit the option to specify the
meeting start/end range. The current implementation lacks support for
returning the availabilityView which shows start/end in the given
availability_view_interval as a string.