Move _server_store call into BackendMiddleware outside of resources
New class variable 'need_store' introduced to make 'store' as an option.
Because all methods in resources have some parameters and we have to pass
them while some of them are not needed.
When we wanted to move some methods from 'mailFolders' to 'messages' and
stuff like that, we have encountered with this that we should pass different
parameters for those methods which were related to the result of
_server_store function and it was annoying because we should put some
conditions for them as well.
Then we've seen _server_store function is being called in many resources and
we decided to put it in the BackendMiddleware to have the required 'store'
at the beginning then we can manage to call methods with different parameters
easily because we already have those data at the beginning and there is no need
any extra conditions.