Implement isOrganiser correctly for shared stores/delegates
isOrganiser always returned True regardless of the item was owned by the
authenticated user. isOrganiser is defined as the authenticated user
being the owner of the calendar or the user being a delegate of the
calendar. This requires the authenticated user's store to be available
in the request context to be able to compare the operated store on the
authenticated users store.