Suggested solution to 'yielder' problem in resources.py
The yielder() took away the advantages of a generator: generate the data in place. By the yielder, the program had to go through the whole generator function before knowing the pagination. This became problematic for huge data sets (i.e. >15000). With the proposed design, the actual query is delivered alongside with the store to the generator function and the generator is fed with the restriction. The restriction is necessary, as only folder.items has a query parameter, whereas folder.folders and folder.items both have a restriction parameter.